+++ weblate/utils/validators.py
@@ -48,19 +48,6 @@ ALLOWED_IMAGES = frozenset((
     'image/apng',
 ))
 
-# List of schemes not allowed in editor URL
-# This list is not intededed to be complete, just block
-# the possibly dangerous ones.
-FORBIDDEN_URL_SCHEMES = frozenset((
-    'javascript',
-    'data',
-    'vbscript',
-    'mailto',
-    'ftp',
-    'sms',
-    'tel',
-))
-
 # File formats we do not accept on translation/glossary upload
 FORBIDDEN_EXTENSIONS = frozenset((
     '.png',
@@ -151,44 +138,6 @@ def validate_bitmap(value):
         raise ValidationError(
             _('Image is too big, please scale it down or crop relevant part!')
         )
-
-
-def validate_repoweb(val):
-    """Validate whether URL for repository browser is valid.
-
-    It checks whether it can be filled in using format string.
-    """
-    try:
-        val % {
-            'file': 'file.po',
-            '../file': 'file.po',
-            '../../file': 'file.po',
-            '../../../file': 'file.po',
-            'line': '9',
-            'branch': 'master'
-        }
-    except Exception as error:
-        raise ValidationError(_('Bad format string (%s)') % str(error))
-
-
-def validate_editor(val):
-    """Validate URL for custom editor link.
-
-    - Check whether it correctly uses format strings.
-    - Check whether scheme is sane.
-    """
-    if not val:
-        return
-    validate_repoweb(val)
-
-    if ':' not in val:
-        raise ValidationError(_('The editor link lacks URL scheme!'))
-
-    scheme = val.split(':', 1)[0]
-
-    # Block forbidden schemes as well as format strings
-    if scheme.strip().lower() in FORBIDDEN_URL_SCHEMES or '%' in scheme:
-        raise ValidationError(_('Forbidden URL scheme!'))
 
 
 def clean_fullname(val):